+++ src/main/java/kr/or/kosa/snippets/like/SnippetsApplication.java
@@ -1,11 +1,11 @@
-package kr.or.kosa.snippets;
+package kr.or.kosa.snippets.like;
 
 import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-@MapperScan("kr.or.kosa.snippets.mapper")
+@MapperScan("kr.or.kosa.snippets.like.mapper")
 public class SnippetsApplication {
 
     public static void main(String[] args) {
