+++ src/main/resources/application.properties
@@ -19,7 +19,7 @@ spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect
 
 # MyBatis ??
 mybatis.mapper-locations=classpath:mapper/*.xml
-mybatis.type-aliases-package=kr.or.kosa
+mybatis.type-aliases-package=kr.or.kosa.snippets.like
 mybatis.configuration.map-underscore-to-camel-case=true
 mybatis.configuration.cache-enabled=false
 mybatis.configuration.jdbc-type-for-null=NULL
