+++ src/main/resources/templates/like/snippet-detail.html
